added main, enabling choice from user

+++ include/constants.hpp
@@ -2,6 +2,6 @@
 
 namespace nbody::constants {
 
-inline constexpr double G = 6.674e-11;
+inline constexpr float G = 6.674e-11;
 inline constexpr float soft = 0.035f;
 }  // namespace nbody::constants
